requirements.txt

fastapi==0.104.1
httpx==0.25.2
uvicorn[standard]==0.24.0
openai>=1.6.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic[email]==2.9.0
python-multipart==0.0.6

# Database
asyncpg>=0.29.0
pgvector>=0.2.4

# Azure Storage
azure-storage-blob>=12.19.0

# Email
sendgrid>=6.10.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# LangGraph dependencies
langgraph
langchain
langchain-openai
langchain-core

# Document processing
docling>=2.5.0
